add some tests

+++ lib/cannon_rabbit/float.rb
@@ -1,4 +1,4 @@
-  class Numeric
+  class Float
     def get_radians
       self * Math::PI / 180 
     end
